+++ src/dacs/ofxLaserDacManagerHelios.cpp
@@ -84,7 +84,7 @@ vector<DacData> DacManagerHelios :: updateDacList(){
 
     
     
-    libusb_free_device_list(libusb_device_list, cnt);
+    libusb_free_device_list(libusb_device_list, (int)cnt);
 
     return daclist;
     
@@ -131,7 +131,7 @@ DacBase* DacManagerHelios :: getAndConnectToDac(const string& id){
         }
     }
 
-    libusb_free_device_list(libusb_device_list, cnt);
+    libusb_free_device_list(libusb_device_list, (int)cnt);
